A couple of changes on the ming32 make bash file.

+++ projects/mingw32/make.sh
@@ -2,7 +2,15 @@
 cd $(dirname "$0")
 premake4 --file=../../premake4.lua --os=windows --platform=mingw32 --with-static-freetype gmake
 cd ../../make/mingw32/
-time make $@ -e verbose=yes
+make $@ -e verbose=yes
 cd ../../libs/mingw32/
-mv eepp.dll ../../
-mv eepp-debug.dll ../../
+
+if [ -f eepp.dll ];
+then
+	mv eepp.dll ../../
+fi
+
+if [ -f eepp-debug.dll ];
+then
+	mv eepp-debug.dll ../../
+fi
